fix(backup): Better sorting for sorted_dependencies() (#58023)
Previously, the sorting between sibling dependencies (that is,
dependencies which could be swapped in the final sorted list without
affecting he validity of the topographic sort) was a bit haphazard, and
could be thrown of by faraway changes like import order. We now do more
alphabetization when sorting, so this is no longer an issue.

The user-facing message that tests raise when the dependency graph has
been changed has also been improved.