feat(issue-stream): Add "unhandled" as collapsable parameter (#57951)
this pr adds in "unhandled" as a collapsable parameter for the issues
stream request. right now, the issue stream request has to wait for the
unhandled stats to be found, which slows things down. to get around
this, we will collapse "unhandled" and make a separate request to find
the "unhandled" information , similar to how we treat "stats".

closes https://github.com/getsentry/sentry/issues/57844