ref(metrics_extraction): Port metrics extraction to use *Builder functions (#58187)
This splits out some shared code from #58108 that ports the metrics
extraction code from using `fields.py` for function parsing and instead
using the empty UnresolvedQueryBuilder. This helps us not have to
reimplement classes in `fields.py` that are already defined in the
querybuilder, as well as have parity on any other updates to discover /
dashboard behaviour.