fix(meps): Only overwrite status code if int
When we fetch transaction.status results from metrics, the status code
that's returned back is not an int but rather a string with the value
we'd typically get if we resolve through the map.

This checks if the value is an int and if it is, fetch from the map,
otherwise don't overwrite it.