feat(suspect-commits): Add GitHub implementation for getting commit context from all frames (#58088)
Closes https://github.com/getsentry/sentry/issues/57433

Note that while this adds funcionality to the GitHub integration, this
code path is only used in tests so far. 

- Adds `get_commit_context_all_frames` and `get_blame_for_files` to
`CommitContextMixin`, which are similar to the existing
`get_commit_context` and `get_blame_for_file`, except that they will
make a request for multiple frames at once
- Implements these functions for GitHub (GitLab will come later)
- Using the GraphQL API, queries are dynamically generated to fetch all
files at once. There is quite a bit of logic in `github/blame.py` which
deals with deduping the list of file info, generating the proper query,
and then extracting/formatting the query data.