fix(stat-detectors): Disable Replays for Duration Regression issues
Disabling the tab for now because the changes to the replay-related
hooks aren't semantic and trivial. e.g. useReplaysCount accepts
transaction names, but we can't apply other filters such as duration
like we do in the Sampled Events tab. The hooks are also hardcoded
to a 14d stats period which doesn't work with our issue data