fix(perf-issues): Remove query from similar spans link
This removes the query from the 'similar spans' link when the button is in an embedded tree. In this case it's embedded inside
performance issues, which have issue filters in their query. In the long term we should likely diverge the query key names or
have context aware parsing of the query when the user switches between areas of the product.