ci: Remove changes outside codecov_ats.yml
Changes introduced to `backend.yml` and the `test-python-ci` target
are being removed.

This is because there was evidence that using `--cov-context=test` in pytest
would incur a 22% (median) increase in CI time. That was deemed unacceptable.

However Codecov's Automated Test Selection still needs this data to be able
to correlate test execution and coverage information with the test that
generated that coverage.

The agreed upon solution will be a periodic job to collect coverage info in
the master branch and a cheaper job to carryforward coverage reports for
all new commits in the master branch.

These changes will be made in a separate PR.