feat(backup): Enable imports over RPC
This finishes out the work started in #57740 and enables imports over
RPC as well. Like that PR, imports are already done on a sequential, per
model basis, so this change just consists of moving every such call
across an RPC boundary.

This is a second attempt at landing this PR, with fewer tests tested in
both modes.