feat(eventuser): Query Snuba for EventUser data (#58124)
## Objective:

We want to confirm that the data in EventUser can be replicated by
either Snuba or the Event. This PR creates a new feature flag. Under
this flag, we query Snuba and then do a 3 way comparison with the
results, the EventUser record and the Event.
We will create a new analytic event so that we can query the results.
And we will also add a log to debug specific events.