feat(stat-detectors): Add hook to query replays for issue samples (#58379)
Adds a hook that queries replays for issues relevant for statistical
detectors. The intention for this is to create a base component for the
replay table and branch off the querying logic to this hook if the group
is a statistical detector issue.

The code is mostly the same as the `useReplaysFromIssue` hook, but some
properties have been hardcoded since this should only be used for
statistical detectors, i.e. perf issues.