test(hc): Change assertions on tests to proxied integration API (#58446)
Introduce IntegratedApiTestCase as a test case superclass with a utility
method for checking whether proxying to the control silo is expected.

In unit tests for integration, resolve assertion errors (typically
`KeyError: 'authorization'`) on proxied requests to external APIs. The
error occurred because the `responses` utility captures only the proxy
request to the control silo, and the underlying request by the control
silo to the external API (which the test asserts for) never happens.

This solution depends on the assumption that the "proxy to control silo"
behavior is basically correct as-is, and is adequately tested elsewhere,
so that we only need to change the assertions to match the existing
behavior. If so, then the monolith-mode test should be enough to test
the correctness of the requests made to the external API. In region
mode, we test only the proxy request to the control silo, and treat the
control silo's behavior as a black box.