feat(issue-stream): Update how unhandled is fetched (#58478)
this pr updates how the unhandled information for groups is handled.
Now, we are collapsing "unhandled" which will no longer find the
unhandled information during the issues request. instead, it will gather
it in the stats request and populate it from there. It is hidden behind
a feature flag, so even though "unhandled" is always being collapsed,
the behavior will only be different if you have the flag.

https://github.com/getsentry/sentry/issues/57847