feat(indexer): Pass use case id along with every message for COGS
The previous approach (where use cases were pre-aggregated in the batching
step was not right. The reason is we can't unbatch it later, which needs to
happen in the step before we actually record the cogs data. The usageaccumulator
library already already does batching internally before producing the messages
so we don't need to worry about calling record() too frequently, and can just
do so once per message.