feat(getting-started): Add Astro onboarding (#58435)
This change adds a getting started page for Astro. Furthermore it also
re-adds astro to the platform picker.

One noteable thing here: We're currently experimenting with enabeling
performance and Replay by default and letting users opt out of both by
specifying 0-sample rates. Therefore, the init code is simplest when
everything is selected and additional options (the rates) are added, if
products are un-selected.