feat(hybridcloud) Add relay endpoints to region pinned list (#58558)
We have several endpoints for configuring new relays and for relay to
request data from sentry with. These endpoints are region bound, but
have no organization scoping available.

Typically we would add `organization_slug` to paths like this to allow
them to be proxied by control silo. I think that is the wrong approach
for these endpoints as these endpoints are either only usable by
superuser or by a relay instance. Relay instances would need to point to
the region domain in order for ingestion to work, and superuser
endpoints don't require as much backwards compatibility as customer
endpoints do.

Binding requests to endpoints on the control silo to the monolith region
is necesary for backwards compatibility.

Refs HC-507
Refs HC-508
Refs HC-509
Refs HC-510