ref(participants): Fix reassignment notifications (#58544)
A follow up to https://github.com/getsentry/sentry/pull/57953 which had
issues not sending the re-assignment notification to the previous
assignee and double sending the assignment notification to the new
assignee. Instead of calling `deassign` ad then `assign` for
reassignment, I moved most of the logic to `assign` which handles
reassignment. I also beefed up the tests by checking the `Activity`
models to make sure those are correct since they are responsible for
sending the notifications.