fix(actionable-items): Fix bug where nothing is shown (#58549)
this pr fixes a problem where the alert would be shown with no errors,
just the header. The bug came from not having an additional check after
a filter, meaning it could think that there were errors to show, but
there weren't. In this scenario, it would show the alert, but then now
show any errors. This is fixed by checking if there are any errors,
after filtering, and if there aren't, not showing the alert.

Fixes https://github.com/getsentry/sentry/issues/58548