ref(auth): check_auth less bursty, does not refresh no op providers (#58563)
1. check_auth runs a single job per auth identity needing processing
every minute it finds groups of auth that need processing. This tends to
clump the processing quite a bit and result in a bursts of outboxes and
processing that pressure the database and celery workers. We instead
process these now serially inside of a job that recursively processes
batches one at a time.

2. check_auth currently does a lot of work updating and processing
outboxes for auth providers that have no refresh logic associated with
it merely to keep the `last_verified` attribute up to date. However, for
some auth providers, such as SAML, this verification process is entirely
a no-op anyways. Change check_auth to ignore those auth identities and
produce far less noise in outbox and check_auth processing.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>