fix(crons): Correctly compute next_checkin for tz aware monitors (#58550)
After the changes made in GH-55874 we introduced a regression that
caused the computation of the `next_checkin` and `next_checkin_latest`
to be incorrect for timezone aware monitors

We were not providing `get_prev_schedule` timezone aware datetime
objects in the timezone of the monitor. Meaning when we would compute
the previous expected check-in time from the current clock tick, we were
computing the schedule in UTC.

An example is as follows

monitor-1 (`0 0 * * *`) -- UTC-5 timezone (America/New_York)

At `01-20-2023 00:00 UTC-5` (`01-20-2023 05:00 UTC+0`) it is expected to
check in

At `01-20-2023 00:00 UTC-5` it does not check in
At `01-20-2023 00:01 UTC-5` it does not check in (missed margin 1)

We now compute when the next expected check in should be. We would
expect it to be `01-21-2023 00:00 UTC-5`.

What we're actually doing though is first computing from `00:01` when
the previous check-in was and using that as the reference time for when
the next check-in should be.

When we were computing the previous check-in time the datetime's
provided to `get_prev_schedule` were NOT timezone aware. This meant when
we computed the schedule of `0 0 * * *` we would get a datetime without
a timezone.

```
get_prev_schedule('01-20-2023 05:00')
```

(note the missing timezone) would return `01-20-2023 00:00`, again
without a timezone. Now we use this to compute the `next_checkin`. The
problem is that when computing the next check in we first translate the
timestamp passed into the monitors timezone, which means
get_next_schedule is called like this

```
get_next_schedule('01-19-2023 19:00 UTC-5')
```

Since the UTC-5 of `00:00 UTC+0` is `19:00` the day before. Now we just
get the same timestamp back that we just marked missed. This timestamp
will keep being marked as missed and keep being recomputed until the
non-timezone aware datetime is computed into the future.

---

The fix for this is to provide `get_prev_schedule` with timezone aware
date times that are in the timezone of the monitor.

This is equivalent to what `get_next_expected_checkin` does on the
Monitor itself.