feat(suspect-commits): Add rate limit checking for GitHub (#58521)
Closes https://github.com/getsentry/sentry/issues/57430

Before fetching file blames, first checks GitHub for the number of
requests remaining. If it is below a certain threshold, we bail out and
fall back to the release-based suspect commits (the threshold ensures we
have enough requests remaining for other features). This will increment
a metric and create a sentry error so we can quantify how often this
happens.