fix(spike-protection): project settings toggle (#58651)
Spike protection toggle in project setting is always on. I debug this
and looks like the option is not passed to the frontend at all and
frontend assumes "enabled" if options.disable-spike-protection does not
exist. This PR fixes the issue by returning that option as part of the
project serializer.

Before:


https://github.com/getsentry/sentry/assets/132939361/32a07120-b7a3-4128-8a4e-a047f054b00f

After:


https://github.com/getsentry/sentry/assets/132939361/f3704251-ec4b-48bd-8d37-2091cd3fb4f0