ref(hc): Adding team and user to alert rules (#58626)
As part of our plan to deprecate `Actor` model in favor of explicit
columns for `user` and `team`, introduces a new, empty column to the
`AlertRule` model, one of the last usages of `Actor`. Rollout plan looks
something like this:

1.  Add nullable columns for user and team  << We are here.
2. Begin writing to user / team columns (still nullable) alongside actor
in creation code paths. Prevent saves that lack user or team ids.
3. Introduce backfill that writes historical user and team values via
lookup to actor.
4. Run backfill in all environments.
5. Add constraint ensuring that user / team are set.
6. Change read paths to using user / team rather than actor (this likely
involves steps in the backend and frontend)
7.  Change actor to nullable column.
8. Remove actor column from state
9. Drop actor column