fix(statistical-detectors): use DURATION_LIGHT metric for timeseries query
The DURATION_LIGHT metric only extracts the `transaction` tag, making it a
lower cardinality dataset (and therefore faster to query) when other tags
aren't needed. The transaction duration timeseries query only filters on transaction, so switch it to use the faster metric.

Also added a note to the use of DURATION for the top transactions query - that
one also filters on `transaction.op` so it isn't eligible for DURATION_LIGHT (a
mistake made in https://github.com/getsentry/sentry/pull/58480).