fix(statistical-detectors): dedupe organization_ids
Project `organization_id`s extracted from a list of projects to be checked for
regressions can result in a lot of duplicates. All those dupes end up being
serialized as job arguments and also as part of the top transaction names SnQL
query. Deduplicate the organization IDs early to avoid the extra serialization.