Introduce `precision` arg to exactDuration formatter (#58697)
Allows us to specify a minimum duration that we care about for the
`getExactDuration` formatter and truncates the remainder

Use case:
- I'm utilizing the `getExactDuration` formatter, but don't need
millisecond precision. This allows me to pin the formatter to just
seconds, or even minutes if necessary.

How does this differ from simply using `getDuration`?
- `getDuration` rounds the values, while `getExactDuration` does not. So
specifying the minimum duration should _not_ round the value, but should
truncate.