ref(metrics): Automatically infer the granularity and use case ID (#58542)
The granularity and use case ID can be determined based on other
parameters
passed in from the query. Automatically determine them based on that.

Note that the granularity inference is not smart. It simply picks the
largest
possible granularity based on the specified time range and interval. If
the
time range is not rounded to the closest X day/hour/minute then it will
default
to the smallest granularity.