fix(stat-detectors): Move diff to clickhouse and improve sorting (#58728)
There were issues with the way we were doing span analysis before. Some
of the issues were:
- We were limited by a response of 10000 rows to do the analysis in
python, meaning that our data was more inaccurate
- We would serialize a lot of objects, only to dump them in the end

The solution was to move the calculations into Clickhouse. This PR adds
a bunch of data to the query so the response is only what we need, and
we can analyze the full amount of spans. It calculates `tpm_before`,
`tpm_after`, `p95_before` and `p95_after`, which are then used to
calculate the impact of the span before/after for the sort order (i.e.
`tpm_after * p95_after - tpm_before * p95_before`). We've found this to
be more accurate in surfacing spans that aren't just one-offs with a
large p95 delta because they're newly added spans.