fix(suspect-commits): Stop making sentry issues for expected errors (#58726)
The suspect commit errors are extremely noisy,
so this is an attempt at filtering out ones we don't care about.

Changes are:

1. Modifies the GitHub client to add status codes to ApiError, because
GraphQL will always return with a 200 when properly authenticated. This
means parsing the errors array for a rate limit message or adding 404
when data does not exist.
2. Modifies the `except` block to use `logger.warning` for 401, 403,
404, and 429 responses.
3. Replaces the old logic which used `sentry_sdk.capure_exception` and
`logger.error` with a single `logger.exception`.