feat(team-workflow): Remove invalid teams as recipients (#58656)
This PR removes any teams that don't use a valid team provider as
recipients for notifications, falling back to notifying users
individually. Valid team providers are those that can be used to send
notifications to a team. Right now, the only valid team provider is
Slack, so we check that the team has a Slack channel linked.

Closes #58347