perf(ui): Optimize CompactSelect to handle a very large number of options (#58790)
For orgs that have 100k+ environments, CompactSelect was causing the
page to become unresponsive. Profiling pointed to this part of the code
as the main culprit, so I made some simple changes to make it run
quicker. With these changes the dropdown loads and is at least usable
with 70k items. There are probably more things we can do here, and we
should definitely revisit environments soon so we don't have to deal
with these kinds of numbers for a client side search.