Ref(metrics): Refactor batch to retain offset and partition of invalid messages (#58675)
Part of my efforts to add dlq to the sentry metrics consumers, see the
final change here:
https://github.com/getsentry/sentry/pull/57998 

Previous PR in the stack:
https://github.com/getsentry/sentry/pull/58474

Next PR in the stack:
https://github.com/getsentry/sentry/pull/58677

### Overview

- Refactor `batch` to maintain both offsets and partition of an invalid
message, from just maintaining the offset. This is done because we need
to produce both offset and partition to let arroyo move messages into
dlq.
- Separate messages that are invalid and filtered, in 2 separate data
structures. This is done because when we are ready to produce to dlq,
the filtered messages do not end up in the dlq.
- Implements a `_extract_message` function as help to
`_extract_messages` so the process of recovering offset can be cleaner
- Update test