feat(team-workflow): Remove invalid teams as recipients (#58820)
This PR removes any teams that don't use a valid team provider as
recipients for notifications, falling back to notifying users
individually. Valid team providers are those that can be used to send
notifications to a team. Right now, the only valid team provider is
Slack, so we check that the team has a Slack channel linked.

--

This PR re-reverts #58656. 

Once #58807 is merged, which adds flags to `RpcOrganizationMapping`,
this feature can be safely re-merged. Previously, an issue occurred
where flags was accessed on an instance of `RpcOrganizationMapping`, so
the original PR needed to be reverted.