fix(statistical-detectors): Move breakpoint feature flags to the breakpoint task
Having the breakpoint feature flags at the end of the EMA task means that
there's no way to turn off the breakpoint detection stage immediately as it'll
sit in the queue for 12 hours. Moving it into the task means even if the task
has been dispatched, we can stop it from running.