ci: changes to codecov_ats
Change output from envvar to JSON files (as artifacts).
This is to avoid blowing up GHActions by exceeding the maximum output size.

Some runs that have this problem:
* https://github.com/giovanni-guidini/sentry/actions/runs/6654679270
* https://github.com/giovanni-guidini/sentry/actions/runs/6654679270

Runs with the probem fixed:
* https://github.com/giovanni-guidini/sentry/actions/runs/6656840288
* https://github.com/giovanni-guidini/sentry/actions/runs/6656919304

Also adding changes to better surface ATS error in GHA UI and have an artifact
with the real, actual result, for data analysis.
Depends on codecov/engineering-team#711