feat(ci) Add test to detect requirements-dev usage in production code
We've had a few rollbacks in recent history related to usage of
requirements-dev packages in production code. Currently these mistakes
cannot be caught by CI, or cloudbuild and can only be detected once
a deploy has been started

By importing key modules we pull in most of the application which lets
us have some assurance that the application will start.

Alternative to #58695 which used flake8 checks to detect incorrect
imports.