feat(feedback): shim user reports to feedback if flag enabled (#58806)
NOTE: the `_shim_to_feedback` function is a bit jank. This is because
this is still pre-alpha, and schema hasn't been completely ironed out.
Once it is ironed out, I'll rewrite this to use better python typing,
and less dictionary jank

- shims old user reports into Feedback events, for the new User Feedback
feature
- Only runs if the ingestion flag for new Feedbacks is turned on
- Attempts to grab information from the user report, and its associated
event if it exists
- Still missing some fields, but wanted to get this out there & confirm
overall flow works. Can do further analysis on what else we can grab, as
seeing production data will help with this.
- If the shim fails, it will catch the exception and continue gracefully
- We're also now validating the schema inside our `create_feedback`
function, as the current validation does not run in tests which makes it
hard to debug/verify, and this isn't expensive to do.
- Adds tests to the user report endpoint, although this code will also
run in ingestion. I think the coverage here is sufficient as the data in
ingestion will be the same from what I see.