feat(stats-detectors): Add per project ratelimit for ema detection
As a precaution, we want to able to rate limit the number of possible
regressions surfaced per project. This is to prevent too many 14 days queries
from being made in the next step in the analysis. This change introduces an
option `statistical_detectors.ratelimit.ema` that allows us to set a limit for
the number of possible regressions per project. It defaults to -1 meaning no
limit, and any non negative value is the number of possible regressions allowed
per project.