feat(metrics-extraction): Add widget columns to extraction
This adds dashboard widget columns to the MetricSpec's emitted from the relay metric extraction code. Moving forward this should mean any widget with columns should start working with on-demand metrics.

Also in this PR is cardinality checking functionality to block emitting any metrics specs that exceeds a high upper bound. We will also be limiting cardinality to a lower number upon dashboard creation, this is mostly to minimize churn on existing dashboards.