fix(ui) Fix domains used in preloading
Our organization preloading was not using region scoped URLs. This
results in slower requests in the test silo environment, and incorrect
requests being made in dev-ui mode.

For the dev-ui environment, I've moved preloading to happen after
client-config is fetched. This ensures that the necessary context has
been fetched to make requests to the correct location.