fix(integrations): Skip check if default is empty array (#59027)
For ticket action fields, we a check to see if the default value is in
the list of choices. We normally avoid this check by setting the default
to an empty string.

For the GitHub action however, we need to set the default value to an
array. This ends up triggering the check as all arrays are truth-y. This
fix allows us to pass the default array without triggering the check so
the user doesn't see the error below when opening the modal.

### Error when opening modal
<img width="608" alt="Screenshot 2023-10-30 at 11 48 33 AM"
src="https://github.com/getsentry/sentry/assets/67301797/f6f4c05d-16e3-45d0-9888-8ade8282130c">