Always upload Release as Artifact Bundles
We have been turning this feature flag up and it is now fully enabled for some time.

This now removes that feature flag in favor of always creating Artifact Bundles, as we have seen that the feature flag can have a very low number of false negatives.

The rest of the code is kept as is, and the legacy code paths can be removed at a later point in time.