feat: Add API endpoints for handling notification unsubscribes
We currently handle unsubscribe requests for issue workflow, digests and
incident notifications through django views. Adapting these views to
work well in hybrid cloud would require either:

1. More complex deployment setup so that we can link to current frontend
   build assets from region silos.
2. Building several RPC methods to read/write records in the region from
   control.
3. Moving to having the UI for unsubsribe views rendered on control and
   doing reads/writes on region API endpoints.

Given that we slowly working towards less HTML rendered in Django,
I chose to go that route. Option 2 felt worse as we're
trying to limit the number of cross region requests we need to make.
Option 1 adds even more operational and deploy time complexity that we
need to maintain and remember. Where as react + region requests is the
same as the rest of the application.

I've not built an endpoint for unsubscribing from incidents as I'm
unable to find a code path that generates signed links from incidents.

Refs HC-948