ref(metrics): Automatically infer the Dataset (#59118)
If the use case ID of the request is for sessions, then switch the
Dataset to
be metrics.

In theory the code using the layer should know which product it is, and
therefore which Dataset to use. However some of the code using this new
layer
is feature agnostic, and it's easier to add this line here instead of
changing
all the product code.