fix(ui) Send preload requests to region urls v2
Our organization preloading was not using region scoped URLs. This
results in slower requests in the test silo environment, and incorrect
requests being made in dev-ui mode.

For the dev-ui environment, I've moved preloading to happen after
client-config is fetched. This ensures that the necessary context has
been fetched to make requests to the correct location.

The changes in #59043 didn't include `withCredentials` resulting in the
preload requests failing which these changes address.