feat(bug reports): Allow "check all" to really check all items (#59178)
This change allows the "check all" checkbox at the top of the new User
Feedback list to really select all of the feedback items that match the
current query filters (respecting archived or not, timeframe, and
anything typed into a searchbox)

What this means is that, you can have 200 feedbacks in total, but only
15 loaded into the browser. When you click the checkbox it'll say "200
selected". If you then click "Archive" or "Mark as read" all 200 will be
updated. This includes optimistic updates, and it'll trigger background
refreshes, just as if a single item were updated. All the goodies.

It might overfetch right now though, because the list cache doesnt take
into account what's visible, it just clears the cache and marks
everything for refetch. Something to follow up on.

Note:
- If all the ids are loaded into the browser (because you've scrolled
down for a while) then every checkbox will be clickable. So you can
select all, but then deselect one, marking 199 as read. I think there's
an optimization to be made if all 200 are in the browser, we don't need
to send 200 ids to the backend; something to follow up on.
- If not all the ids are loaded, and you choose to 'select all'. Then
the checkboxes for each item will become disabled. In this state you
cannot uncheck a specific feedback, you need to uncheck all first.