fix(notifications): HC bug serializing external provider (#59066)
We didn't have region silo tests enabled for certain notification
participant tests. In the future, I will ensure all the tests run with
the region in a follow up PR but for now I'm setting just for one to
make sure that the interface logic is correct. This PR fixes a problem
where we couldn't serialize the response for that uses
`ExternalProvider` as a dictionary key instead of a string. This PR uses
a string then converts it back to an enum. In the future, we'll try to
clean up the enum code once we can rip out the code for old notification
work.