feat(metrics): Add option to disable schema validation (#58666)
Add ability to have fine grained control on whether to apply schema
validation checks or not. This is primarily targeted for the generic
metrics pipeline where we want to have more processing power. This PR
does not change anything in the processing pipeline right now. That
would happen in a separate PR.