tests(hybrid-cloud): Stabilize access log middleware tests (#59193)
Stabilizes tests for the api access log middleware. 

Added a test case for the ControlSilo Organization endpoint as well
which showed a failure since `request._request.organization` wasn't set.
Setting that to the RPC organization resolved this but mentionning it
specifically in case that could cause unexpected behaviour down the line
by setting that attr.