feat(issue-stream): Find search from backend  (#59127)
this pr allows the main issues request from the issue stream to be made
without waiting for the saved searches request. 2 additional parameters
are optionally passed to the backend: one if there is a specific saved
search and one if saved searches have not loaded from the front end. If
saved searched are not loaded from the front end, it will check if there
is a specific saved search requested. If there is, it will be used. If
not, it will check if the user has a default search set, and use it if
it is found. If not, it will used the default "is:unresolved" query.

Closes https://github.com/getsentry/sentry/issues/57853