fix(hybrid-cloud): Make wizard interface silo-safe (#59153)
This PR fixes the silo boundaries crossed by the wizard. I opted to put
this on the control silo since moving it to a region would necessitate
changes on the wizard CLI, plus it could make auth tokens which is a
control operation.

Doing so meant that it has to fanout to identify which organizations the
user belongs to and request projects + DSNs from those regions. For now,
to maintain behaviour, it will do the fan out, but I've left a comment
in code and I'll reiterate here that the CLI should display the region
or ask the user which region they're wizarding since project slugs
aren't unique globally. The response also no longer serializes the whole
`Organization`, `Project` and `ProjectKey` objects for the user [since
exploring the code here](https://github.dev/getsentry/sentry-wizard)
shows that only the few fields I provided are actually accessed, so we
can use OrganizationMappings and minimal dictionaries for project and
keys.

I thought about formalizing JSON serializing the RPC objects but haven't
at the moment since neither implemented `FilterQueryDatabaseImpl` which
would have made it easier to standardize