feat(hybridcloud) Update unsubscribe links to use react views
Implement new signed link generation that generates links that contain
either a customer domain path, or a slug based path. The signature data
is for the API endpoint used by the react views.

I've used an option so that I can test this in the test-silos before
enabling it in saas.