feat(issue-stream): Update issue request to not rely on saved search (#59226)
this pr builds off of https://github.com/getsentry/sentry/pull/59127
where the backend is updated to no longer require saved searches to be
fetched first. now, the front end will not have to wait on saved
searches to make the first issues request. after the first time, saved
searches are cached and can be used.

closes https://github.com/getsentry/sentry/issues/57854