ref(rules): Remove unused NeglectedRule rows (#59367)
We ran a 2nd batch of rules to become disabled, but decided not to go
forward with it for now. If we want to pick it back up again, due to the
nature of the tasks that run for the feature, we need to remove the rows
that went through the first two tasks (table population and initial
email sending) to ensure the third task (disabling and sending a final
email) doesn't get triggered when we add the task scheduling back in
later on.