fix(ui): Correctly show new values when updating issue form (#59334)
There was a previous frontend bug in the external issue modal. Basically
when switching a field value that fetches a new field configurations, we
would carry over the last selected value + options even if they didn't
exist in the fetched configuration.
This only affects Alert Rule ticket actions, because it's not possible
to save values when manually creating an issue, because we don't store
the selected fields and create the external issue immediately.

### Bug in Action
In this example below, before saving the `IS` project has several
`Component/s` + `fixVersions`, but `SEIJI` has none. After we save, the
fields carry over when we switch to `SEIJI` even though they don't exist
in the project.


https://github.com/getsentry/sentry/assets/67301797/2dbd5596-e08d-43ff-a201-e93b141b16d6

### After Fix
`Component/s` + `fixVersions` correctly resets after switching projects.


https://github.com/getsentry/sentry/assets/67301797/2ee8f4ce-5c20-46e0-bb1e-c8d97e7005f1